requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
openai>=1.40.0
python-dotenv>=1.0.1

